Updates / Fixes on AmChart core libraries.
- Fixed params, and removed unused classes.
- Fixed compilation issue.

+++ src/main/java/gwt/material/design/amcharts/client/ui/chart/js/options/AmChartScrollbar.java
@@ -141,7 +141,7 @@ public class AmChartScrollbar {
     public String selectedBackgroundColor;
 
     @JsProperty
-    public int selectedGraphFillAlpha;
+    public double selectedGraphFillAlpha;
 
     @JsProperty
     public String selectedGraphFillColor;
